requirements.txt
opencv-python
Pillow
tk
requests
langchain
langchain-community
gradio